Fix build-tags in libcontainer/devices
Allows importing this package on Windows (for the types)
