+++ devices/device_windows.go
@@ -1,5 +1,5 @@
 package devices
 
-func (d *Rule) Mkdev() (uint64, error) {
+func mkDev(_ *Rule) (uint64, error) {
 	return 0, nil
 }
